use different controllers and remove redundant js
